requirements.txt
matplotlib
numpy
openai
pandas
seaborn
streamlit
python-dotenv
